requirements.txt
# Core Dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Text Processing
nltk>=3.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Data Validation
jsonschema>=4.19.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Async Support
aiofiles>=23.2.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
colorama>=0.4.6

# Optional: For advanced features
# anthropic>=0.7.0  # If using Claude for comparison
# langchain>=0.1.0  # If integrating with LangChain
